+++ Editor/translations/Editor_fr.ts
@@ -164,6 +164,27 @@
     </message>
 </context>
 <context>
+    <name>BlockChipView</name>
+    <message>
+        <location filename="../widgets/routing/BlockChipRoutingRenderer.cpp" line="277"/>
+        <source>Show fewer channels</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message numerus="yes">
+        <location filename="../widgets/routing/BlockChipRoutingRenderer.cpp" line="278"/>
+        <source>Show %n more channel(s)</source>
+        <translation type="unfinished">
+            <numerusform></numerusform>
+            <numerusform></numerusform>
+        </translation>
+    </message>
+    <message>
+        <location filename="../widgets/routing/BlockChipRoutingRenderer.cpp" line="291"/>
+        <source>Add channel</source>
+        <translation type="unfinished">Ajouter un canal</translation>
+    </message>
+</context>
+<context>
     <name>ChannelCardEditor</name>
     <message>
         <location filename="../widgets/cards/ChannelCardEditor.cpp" line="21"/>
@@ -1049,7 +1070,7 @@ Supprimer un terme (d&apos;addition)</translatorcomment>
         <translation>Modifier la commande brute</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="581"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="596"/>
         <source>Raw</source>
         <translation>Brut</translation>
     </message>
@@ -2665,7 +2686,7 @@ Voulez-vous exécuter l’application Device Selector (Sélecteur de périphéri
 <context>
     <name>StudioRoutingView</name>
     <message>
-        <location filename="../widgets/routing/LightTraceRoutingRenderer.cpp" line="426"/>
+        <location filename="../widgets/routing/LightTraceRoutingRenderer.cpp" line="585"/>
         <source>Drag between channels to route - click + to add an output</source>
         <translation>Glissez entre les canaux pour router - cliquez sur + pour ajouter une sortie</translation>
     </message>